requirements.txt
jupyterlab
pycairo
